rename update_client to update_state and fix data type
